Restructure files for ESPHome external_components
